chore: remove old core source paths
