+++ crates/attestation/src/platforms/snp/attest.rs
@@ -1,3 +1,5 @@
+use std::path::Path;
+
 use base64::{engine::general_purpose::STANDARD as BASE64, Engine};
 
 use sev::firmware::guest::Firmware;
@@ -8,11 +10,12 @@ use crate::utils::pad_report_data;
 
 use super::evidence::{SnpCertChain, SnpEvidence};
 
-const SEV_PLATFORM_PATH: &str = "/sys/devices/platform/sev-guest";
+/// SNP guest attestation character device.
+const SEV_GUEST_DEVICE_PATH: &str = "/dev/sev-guest";
 
 /// Check if SNP hardware is available on this machine.
 pub fn is_available() -> bool {
-    std::path::Path::new(SEV_PLATFORM_PATH).exists()
+    Path::new(SEV_GUEST_DEVICE_PATH).exists()
 }
 
 /// Extract certificates from the sev crate's cert table entries.
