--- src/platforms/az_snp/mod.rs
@@ -1,5 +1,0 @@
-pub mod evidence;
-pub mod verify;
-
-#[cfg(all(feature = "attest", feature = "az-snp", target_os = "linux"))]
-pub mod attest;
