--- src/platforms/az_tdx/mod.rs
@@ -1,5 +1,0 @@
-pub mod evidence;
-pub mod verify;
-
-#[cfg(all(feature = "attest", feature = "az-tdx", target_os = "linux"))]
-pub mod attest;
