--- src/platforms/dstack/mod.rs
@@ -1,5 +1,0 @@
-pub mod evidence;
-pub mod verify;
-
-#[cfg(all(feature = "attest", target_os = "linux"))]
-pub mod attest;
