--- src/platforms/gcp_snp/mod.rs
@@ -1,4 +1,0 @@
-pub mod verify;
-
-#[cfg(all(feature = "attest", feature = "gcp-snp", target_os = "linux"))]
-pub mod attest;
