--- src/platforms/gcp_tdx/mod.rs
@@ -1,4 +1,0 @@
-pub mod verify;
-
-#[cfg(all(feature = "attest", feature = "gcp-tdx", target_os = "linux"))]
-pub mod attest;
