--- src/platforms/mod.rs
@@ -1,16 +1,0 @@
-#[cfg(feature = "az-snp")]
-pub mod az_snp;
-#[cfg(feature = "az-tdx")]
-pub mod az_tdx;
-#[cfg(feature = "dstack")]
-pub mod dstack;
-#[cfg(feature = "gcp-snp")]
-pub mod gcp_snp;
-#[cfg(feature = "gcp-tdx")]
-pub mod gcp_tdx;
-#[cfg(feature = "snp")]
-pub mod snp;
-#[cfg(feature = "tdx")]
-pub mod tdx;
-#[cfg(any(feature = "az-snp", feature = "az-tdx"))]
-pub mod tpm_common;
