--- src/platforms/snp/mod.rs
@@ -1,7 +1,0 @@
-pub mod certs;
-pub mod claims;
-pub mod evidence;
-pub mod verify;
-
-#[cfg(all(feature = "attest", feature = "snp", target_os = "linux"))]
-pub mod attest;
