--- src/platforms/tdx/mod.rs
@@ -1,8 +1,0 @@
-pub mod ccel;
-pub mod claims;
-pub mod dcap;
-pub mod evidence;
-pub mod verify;
-
-#[cfg(all(feature = "attest", feature = "tdx", target_os = "linux"))]
-pub mod attest;
